Fix foundation.app.handle otel span
Currently the span is being created before there's a tracer in the
context. As a result, the AddSpan only creates a new span, without
actually starting it and attaching the received attributes.
So, the solution is to construct the span and ctx before calling the
AddSpan (similar to what happened before).

I also added an endpoint attribute, just thought it could be helpful.